created the random bubble generator
random bubble generator to implement color checks. the colorcheck function still needs to be linked to the random color generator.
Plus changes in the instantiation of the director bubble(where the colored bubbles show up) needs to be made such that it get instantiated with the board and the colored bubbles start showing up after some time. (will need it when I start implementing game states of play, refill and pause)